--- src/apps/vanilla/constants.ts
@@ -1,3 +1,0 @@
-import type { TasksView } from '@/apps/vanilla/types'
-
-export const allViews: TasksView[] = ['list', 'detailed', 'condensed']
