+++ src/apps/vanilla/context/TasksContext.tsx
@@ -1,5 +1,5 @@
 import { createContext } from 'react'
 
-import type { TasksState } from '@/apps/vanilla/types'
+import type { TasksState } from '@/apps/common/types'
 
 export const TasksContext = createContext<TasksState>(null as unknown as TasksState)
