--- src/apps/zustand/constants.ts
@@ -1,3 +1,0 @@
-import type { TasksView } from '@/apps/zustand/types'
-
-export const allViews: TasksView[] = ['list', 'detailed', 'condensed']
